Add build script
