+++ config/index.js
@@ -23,7 +23,7 @@ import {
 } from './names'
 import envConfig from './env.config'
 import networkConfig from './network.config'
-import manifest from '../extension/manifest_template.json'
+import manifest from './manifest_template.json'
 import dotenv from 'dotenv'
 dotenv.config({})
