+++ config/manifest_template.json
@@ -1,0 +1,53 @@
+{
+  "manifest_version": 2,
+  "name": null,
+  "version": "0.5.2",
+  "description": "",
+  "browser_action": {
+    "default_title": null,
+    "default_popup": "index.html"
+  },
+  "author": "Gnosis",
+  "default_locale": "en",
+  "options_ui": {
+    "page": "options.html",
+    "open_in_tab": true
+  },
+  "icons": {
+    "16": "assets/images/favicon.png",
+    "48": "assets/images/favicon.png"
+  },
+  "permissions": [
+    "storage",
+    "activeTab",
+    "tabs",
+    "notifications"
+  ],
+  "background": {
+    "scripts": [
+      "background.js"
+    ],
+    "persistent": true
+  },
+  "content_scripts": [
+    {
+      "matches": [
+        "http://*/*",
+        "https://*/*"
+      ],
+      "js": [
+        "contentscript.js"
+      ],
+      "run_at": "document_start",
+      "all_frames": true
+    }
+  ],
+  "web_accessible_resources": [
+    "script.js"
+  ],
+  "content_security_policy": "script-src 'self' https://ssl.google-analytics.com https://cdn.firebase.com https://*.firebaseio.com https://apis.google.com; object-src 'self'",
+  "externally_connectable": {
+    "ids": [],
+    "matches": []
+  }
+}
