+++ webpack.config.babel.js
@@ -5,29 +5,11 @@ const CopyWebpackPlugin = require('copy-webpack-plugin')
 const ExtractTextPlugin = require('extract-text-webpack-plugin')
 const cssVars = require('postcss-simple-vars')
 const cssVariables = require(path.resolve(__dirname, './app/theme/variables'))
-const writeJsonFile = require('write-json-file')
-const manifest = require('./extension/manifest_template.json')
 const config = require(path.resolve(__dirname, './config'))
-const names = require(path.resolve(__dirname, './config/names'))
-const dotenv = require('dotenv')
-const env = dotenv.config().parsed
+const build = require(path.resolve(__dirname, './scripts/build'))
 
-const envKeys = Object.keys(env).reduce((prev, next) => {
-  prev[`process.env.${next}`] = JSON.stringify(env[next])
-  return prev
-}, {})
-envKeys['process.env.NODE_ENV'] = process.env.NODE_ENV ? `"${process.env.NODE_ENV}"` : `"${names.DEVELOPMENT}"`
-envKeys['process.env.NETWORK'] = process.env.NETWORK ? `"${process.env.NETWORK}"` : `"${names.RINKEBY}"`
-
-const title = (config.getNetwork() === names.MAINNET)
-  ? 'Gnosis Safe'
-  : 'Gnosis Safe - Rinkeby'
-
-console.log(title)
-
-manifest.name = title
-manifest.browser_action.default_title = title
-writeJsonFile.sync('./build/manifest.json', manifest)
+const envKeys = build.generateEnvVariables()
+build.generateManifestFile()
 
 const postcssPlugins = [
   cssVars({
@@ -135,6 +117,13 @@ module.exports = {
     }),
     new ExtractTextPlugin('css/[name].[contenthash:8].css'),
     new webpack.DefinePlugin(envKeys),
+    new CopyWebpackPlugin([
+      {
+        from: path.resolve(__dirname, './config/_locales'),
+        to: path.resolve(__dirname, './build/_locales'),
+        force: true
+      }
+    ]),
     new CopyWebpackPlugin([
       {
         from: path.resolve(__dirname, './app/assets/images/', config.getFavicon()),
